% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_gene_sample_matrix}
\alias{aggregate_gene_sample_matrix}
\title{Generate Gene-Sample Matrix}
\usage{
aggregate_gene_sample_matrix(normalized_shap_list, absolute)
}
\arguments{
\item{normalized_shap_list}{A named list of normalized SHAP dataframes.}

\item{absolute}{Logical. If TRUE, use absolute values of normalized SHAP; if FALSE, use original values.}
}
\value{
A wide matrix of Gene x Sample.
}
\description{
This function takes a list of normalized SHAP dataframes (one per model) and
creates a wide matrix where rows represent genes and columns represent samples.
The values in the matrix are the mean normalized SHAP values for each gene-sample
combination, averaged across all models. The 'absolute' parameter allows the user
to choose whether to use the absolute values of the normalized SHAP values or to
keep their original sign, depending on whether they want to focus on the magnitude
of importance or also consider the direction of the effect.
}
