% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_genes}
\alias{aggregate_genes}
\title{Generate Gene-Level Aggregation}
\usage{
aggregate_genes(normalized_shap_list)
}
\arguments{
\item{normalized_shap_list}{A named list of normalized SHAP dataframes.}
}
\value{
A summary dataframe with metrics per gene.
}
\description{
This function takes a list of normalized SHAP dataframes (one per model) and
aggregates the importance of each gene across all models. It calculates several
metrics for each gene, including the maximum, mean, median, standard deviation of
the mean normalized SHAP values, and the count of how many models included that gene.
The resulting summary dataframe is sorted by the maximum aggregated importance, allowing
for easy identification of top candidate genes.
}
