% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barplot_feature_selection}
\alias{barplot_feature_selection}
\title{Plot Barplot of input lists}
\usage{
barplot_feature_selection(feature_selection_data, output_file = NULL)
}
\arguments{
\item{feature_selection_data}{Dataframe with columns 'input_list' and 'intersection_count'.}

\item{output_file}{Character. Path to save the output PDF.}
}
\value{
Invisibly returns the plot object.
}
\description{
This function generates a barplot showing the number of genes in each input list used for feature selection. It uses the
package-wide color scheme for consistency.
}
