% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainability.R
\name{calculate_all_shap}
\alias{calculate_all_shap}
\title{Calculate SHAP values for all ML models}
\usage{
calculate_all_shap(
  ml_models_to_run,
  results_models,
  expression_data,
  target_var,
  directory_to_save = NULL
)
}
\arguments{
\item{ml_models_to_run}{Character vector of model names.}

\item{results_models}{List of fitted models (output from ML_models).}

\item{expression_data}{Dataframe. The data used for training.}

\item{target_var}{Character. The target variable column name.}

\item{directory_to_save}{Character. Optional directory path.}
}
\value{
A list containing SHAP values for each model and the time taken for the entire process.
}
\description{
This function iterates over a list of machine learning models, calculates SHAP values for each model using the
\code{calculate_variable_importance_batch} function, and compiles the results into a list. It also ensures that
the necessary package for explaining tidymodels is available and handles the renaming of model labels for better
readability in the output.
}
