% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calculate_sample_clustering}
\alias{calculate_sample_clustering}
\title{Perform Sample Clustering}
\usage{
calculate_sample_clustering(mat, method = "cosine", k = 5)
}
\arguments{
\item{mat}{Matrix. Expression or SHAP matrix (genes as rows, samples as columns).}

\item{method}{Character. "euclidean" or "cosine".}

\item{k}{Integer. Number of clusters to cut.}
}
\value{
A list containing the hierarchical clustering object and the cluster assignments.
}
\description{
This function computes a distance matrix based on the provided method
(either "euclidean" or "cosine") and performs hierarchical clustering
on the samples. It then cuts the resulting dendrogram into a specified
number of clusters (k) and returns both the hierarchical clustering object
and the cluster assignments for each sample.
}
