% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_analysis}
\alias{dea_analysis}
\title{Post-DEA Analysis and Diagnostics}
\usage{
dea_analysis(fit_object, pv = 0.05, lfc = 0, coef = 1, output_dir)
}
\arguments{
\item{fit_object}{MArrayLM object. Output from \code{run_dea_limma}.}

\item{pv}{Numeric. P-value threshold.}

\item{lfc}{Numeric. Log-fold change threshold.}

\item{coef}{Integer. Coefficient to analyze.}

\item{output_dir}{Character. Directory to save plots.}
}
\value{
The results Matrix with DEGs.
}
\description{
This function takes the fitted model object from the DEA analysis and performs post-analysis diagnostics.
It generates several plots to visualize the results, including a Venn diagram of DEGs, a histogram of p-values,
and a mean-difference (MD) plot. The function also returns a matrix indicating which genes are differentially
expressed based on the specified p-value and log-fold change thresholds. The generated plots are saved in a
"DEA_plots" subdirectory within the specified output directory.
}
