% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_ranking.R
\name{get_candidate_genes}
\alias{get_candidate_genes}
\title{Select Candidate Genes by Threshold}
\usage{
get_candidate_genes(analysis_list, threshold)
}
\arguments{
\item{analysis_list}{List. The output from \code{compute_density_analysis}.}

\item{threshold}{Numeric. The threshold decided by the user.}
}
\value{
A sorted dataframe of selected genes.
}
\description{
This function filters the original data based on a user-defined threshold for the specified column.
It returns a sorted dataframe of selected genes and prints the number and percentage of genes selected.
}
