% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_phenotype.R
\name{get_enrichment_logical_matrix}
\alias{get_enrichment_logical_matrix}
\title{Create Enrichment Logical Matrix}
\usage{
get_enrichment_logical_matrix(enrichment_df, alpha = 0.05)
}
\arguments{
\item{enrichment_df}{Output from \code{run_phenotype_enrichment}.}

\item{alpha}{Numeric. Significance threshold.}
}
\value{
A logical matrix indicating significant gene-phenotype associations.
}
\description{
This function takes the output from \code{run_phenotype_enrichment} and creates a logical
matrix indicating which genes are significantly associated with each phenotypic variable
based on adjusted p-values. The resulting matrix has genes as rows and phenotypic variables
as columns, with TRUE indicating significant association and FALSE otherwise.
}
