% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_phenotype.R
\name{get_enrichment_matrix}
\alias{get_enrichment_matrix}
\title{Create Enrichment Matrix padjusted values}
\usage{
get_enrichment_matrix(enrichment_df, alpha = 0.05)
}
\arguments{
\item{enrichment_df}{Output from \code{run_phenotype_enrichment}.}

\item{alpha}{Numeric. Significance threshold.}
}
\value{
A matrix of adjusted p-values for gene-phenotype associations.
}
\description{
This function takes the output from \code{run_phenotype_enrichment} and creates a matrix of
adjusted p-values for each gene-phenotype association. The resulting matrix has genes
as rows and phenotypic variables as columns, with the values representing the adjusted
p-values for each association.
}
