% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{glm_multiclass_model}
\alias{glm_multiclass_model}
\title{GLM Multiclass Model}
\usage{
glm_multiclass_model(
  data,
  target_var,
  folds,
  metrics,
  control,
  directory_to_save = NULL
)
}
\arguments{
\item{data}{Data frame containing the training data.}

\item{target_var}{Character string of the target variable column name.}

\item{folds}{Resampling object (e.g., vfold_cv) for cross-validation}

\item{metrics}{Metrics object for model evaluation.}

\item{control}{Control object for tuning and resampling.}

\item{directory_to_save}{Character string specifying the directory to save the model bundles. Default is NULL (do not save).}
}
\value{
A list containing the time taken for training, resampling results, and the fitted model object.
}
\description{
This function trains a Generalized Linear Model (GLM) for multiclass classification using the
\code{parsnip} and \code{workflows} packages. It performs hyperparameter tuning using Bayesian optimization
and evaluates the model using resampling. The fitted model and resampling results can be saved
as bundles in a specified directory.
}
