% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{gtex_gene_map}
\alias{gtex_gene_map}
\title{Default Gene Name Mapping for GTEx}
\format{
An object of class \code{character} of length 11.
}
\usage{
gtex_gene_map
}
\description{
Default Gene Name Mapping for GTEx
}
\keyword{internal}
