% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{load_gene_list}
\alias{load_gene_list}
\title{Load a gene list from a file}
\usage{
load_gene_list(genes_list, base_path = "../COPD/feature_selection/")
}
\arguments{
\item{genes_list}{Character string of the filename.}

\item{base_path}{Base directory. Defaults to "../COPD/raw_data/".}
}
\value{
A character vector of gene names.
}
\description{
This function loads a gene list from a specified text file. The file is expected to
contain gene names separated by commas.
}
