% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{obtain_alternative_genes}
\alias{obtain_alternative_genes}
\title{Obtain alternative selection of genes}
\usage{
obtain_alternative_genes(
  alternative_genes,
  directory_to_load,
  directory_to_save = NULL
)
}
\arguments{
\item{alternative_genes}{Character string of the filename containing the alternative gene list.}

\item{directory_to_load}{Directory to load the gene list from.}

\item{directory_to_save}{Directory to save the loaded gene list to (optional).}
}
\value{
A character vector of gene names loaded from the specified file.
}
\description{
This function loads an alternative list of genes from a specified text file. The file is expected to
contain gene names separated by commas. The function also saves the loaded gene list to a specified
directory if the \code{directory_to_save} parameter is provided.
}
