% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{obtain_data_driven_genes}
\alias{obtain_data_driven_genes}
\title{Obtain Data Driven Genes (DEA + mRMR)}
\usage{
obtain_data_driven_genes(
  dea_genes = NULL,
  mrmr_genes = NULL,
  expression_data,
  target_var,
  threshold_value = NULL,
  directory_to_load,
  directory_to_save = NULL,
  mrmr_path = "./mrmr"
)
}
\arguments{
\item{dea_genes}{List of differentially expressed genes (optional).}

\item{mrmr_genes}{List of mRMR selected genes (optional).}

\item{expression_data}{Data frame containing the expression data.}

\item{target_var}{Character string of the target variable column name.}

\item{threshold_value}{Numeric value for the mRMR score threshold. If NULL, bootstrapping will be performed to determine the threshold.}

\item{directory_to_load}{Directory to load the DEA results from.}

\item{directory_to_save}{Directory to save the selected data-driven genes to (optional).}

\item{mrmr_path}{Path to the mRMR executable.}
}
\value{
A character vector of selected data-driven gene names.
}
\description{
This function combines the results from Differential Expression Analysis (DEA) and mRMR feature
selection to obtain a list of data-driven genes. It takes in the lists of DEA and mRMR genes,
as well as the expression data and target variable, to determine the final set of genes.
The resulting list can be saved to a specified directory if the \code{directory_to_save} parameter is provided.
}
