% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{obtain_disease_related_curated_genes}
\alias{obtain_disease_related_curated_genes}
\title{Obtain curated disease genes}
\usage{
obtain_disease_related_curated_genes(
  disease_code,
  directory_to_load,
  directory_to_save = NULL
)
}
\arguments{
\item{disease_code}{Disease code for which to obtain the curated gene list.}

\item{directory_to_load}{Directory to load the gene list from.}

\item{directory_to_save}{Directory to save the loaded gene list to (optional).}
}
\value{
A character vector of curated gene symbols related to the specified disease.
}
\description{
This function loads a curated list of disease-related genes from a specified file. The file is expected
to be a TSV file containing gene-disease associations with a "curated" curation effort for the given disease code.
The function extracts the gene symbols from the file and saves the list to a specified directory if the
\code{directory_to_save} parameter is provided.
}
