% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{obtain_disease_related_entire_genes}
\alias{obtain_disease_related_entire_genes}
\title{Obtain disease-related entire genes}
\usage{
obtain_disease_related_entire_genes(
  disease_code,
  directory_to_load,
  directory_to_save = NULL
)
}
\arguments{
\item{disease_code}{Disease code for which to obtain the gene list.}

\item{directory_to_load}{Directory to load the gene list from.}

\item{directory_to_save}{Directory to save the loaded gene list to (optional).}
}
\value{
A character vector of gene names related to the specified disease.
}
\description{
This function loads the entire list of disease-related genes from a specified file. The file is
expected to be a TSV file containing a summary of gene-disease associations for the given disease code.
The function extracts the gene names from the file and saves the list to a specified directory if the
\code{directory_to_save} parameter is provided.
}
