% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{obtain_mrmr}
\alias{obtain_mrmr}
\title{Obtain mRMR genes}
\usage{
obtain_mrmr(
  expression_data,
  target_var,
  threshold_value = NULL,
  directory_to_save = NULL,
  mrmr_path = NULL
)
}
\arguments{
\item{expression_data}{Data frame containing the expression data.}

\item{target_var}{Character string of the target variable column name.}

\item{threshold_value}{Numeric value for the mRMR score threshold. If NULL, bootstrapping will be performed to determine the threshold.}

\item{directory_to_save}{Directory to save the selected mRMR genes to (optional).}

\item{mrmr_path}{Path to the mRMR executable.}
}
\value{
A character vector of selected mRMR gene names.
}
\description{
This function runs the mRMR feature selection algorithm on the provided expression data and target variable.
It can also perform bootstrapping to determine a score threshold for selecting features. The selected mRMR
genes are saved to a specified directory if the \code{directory_to_save} parameter is provided.
}
