% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{obtain_omnipath_combined}
\alias{obtain_omnipath_combined}
\title{Combined Omnipath Expansion (Intersection/Union)}
\usage{
obtain_omnipath_combined(
  expression_data,
  target_var,
  disease_code,
  threshold_value,
  dea_genes,
  mrmr_genes,
  directory_to_load,
  directory_to_save = NULL,
  operation = "intersection",
  mrmr_path = "./mrmr"
)
}
\arguments{
\item{expression_data}{Data frame containing the expression data.}

\item{target_var}{Character string of the target variable column name.}

\item{disease_code}{Disease code for disease-related selections.}

\item{threshold_value}{Numeric value for the mRMR score threshold. If NULL, bootstrapping will be performed to determine the threshold.}

\item{dea_genes}{List of differentially expressed genes (optional).}

\item{mrmr_genes}{List of mRMR selected genes (optional).}

\item{directory_to_load}{Directory to load the necessary data from.}

\item{directory_to_save}{Directory to save the combined gene list to (optional).}

\item{operation}{Character string specifying the combination operation: "intersection" or "union".}

\item{mrmr_path}{Path to the mRMR executable.}
}
\value{
A character vector of combined gene names based on the specified operation.
}
\description{
This function combines the Omnipath expansions of both disease-related and data-driven gene lists
using either intersection or union operations. It takes in the necessary parameters to obtain the
disease-related and data-driven gene lists, performs the Omnipath expansion for each, and then
combines the results based on the specified operation. The final list of genes can be saved to
a specified directory if the \code{directory_to_save} parameter is provided.
}
