% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepreocess.R
\name{obtain_split_data}
\alias{obtain_split_data}
\title{Split Data into Training and Test Sets}
\usage{
obtain_split_data(
  directory_to_load,
  file_name,
  target_var,
  directory_to_save = NULL,
  verbose = TRUE
)
}
\arguments{
\item{directory_to_load}{Directory to load the expression data from.}

\item{file_name}{Name of the expression data .Rda file (without extension).}

\item{target_var}{Name of the target variable in the dataset.}

\item{directory_to_save}{Directory to save the split datasets (optional).}
}
\value{
A list containing the training and test datasets as dataframes.
}
\description{
This function loads an expression dataset, splits it into training and test sets while maintaining
class balance, and optionally saves the resulting datasets as .Rda files. It also prints the class imbalance
percentages for both the training and test sets to the console.
}
