% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust_ranking.R
\name{plot_mclust_diagnostic}
\alias{plot_mclust_diagnostic}
\title{Plot Mclust Diagnostics}
\usage{
plot_mclust_diagnostic(mclust_list, sel_groups = NULL, output_file = NULL)
}
\arguments{
\item{mclust_list}{List. Output from \code{run_mclust_analysis}.}

\item{sel_groups}{Numeric vector. Optional: clusters to highlight with a threshold line.}

\item{output_file}{Character. Optional: path to save as PDF.}
}
\value{
NULL. Generates plots for diagnostics.
}
\description{
Visualizes BIC, ICL and the resulting clusters relative to the SHAP scores.
}
