% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_radarchart}
\alias{plot_radarchart}
\title{Plot Radar Chart}
\usage{
plot_radarchart(data, metric, plot_name, output_dir = NULL, min, max, seq)
}
\arguments{
\item{data}{List. Output from \code{prepare_data_for_radarchart}.}

\item{metric}{Character. Metric name for the title.}

\item{plot_name}{Character. Name for the output file.}

\item{output_dir}{Character. Directory to save the PDF.}

\item{min}{Numeric. Minimum axis value.}

\item{max}{Numeric. Maximum axis value.}

\item{seq}{Numeric. Step for axis labels.}
}
\value{
Invisibly returns the plot object.
}
\description{
Generates a PDF radar chart with specific color logic for known and unknown input lists.
}
