% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_shap_occurrence}
\alias{plot_shap_occurrence}
\title{Plot SHAP Magnitude vs Occurrence}
\usage{
plot_shap_occurrence(
  data,
  y_threshold = NULL,
  title = "SHAP Magnitude vs Occurrence"
)
}
\arguments{
\item{data}{Dataframe containing columns 'occurrence_count', 'aggregated_max', and 'status'.}

\item{y_threshold}{Numeric. Optional threshold for SHAP magnitude to be indicated with a horizontal line.}

\item{title}{Character. Title for the plot.}
}
\value{
A ggplot object with marginal density plots.
}
\description{
This function creates a scatter plot of SHAP magnitude (aggregated_max) against occurrence
count for each gene, colored by gene status. An optional horizontal line can be added to indicate a threshold for SHAP magnitude.
Marginal density plots are included on the axes to show the distribution of occurrence counts and SHAP magnitudes.
}
