% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_message}
\alias{print_message}
\title{Print Timestamped Messages}
\usage{
print_message(..., verbose = getOption("EBEx.verbose", default = TRUE))
}
\arguments{
\item{...}{Arguments to be passed to the message.}

\item{verbose}{Logical. If FALSE, the message is suppressed.
Defaults to the global option "EBEx.verbose" or TRUE.}
}
\description{
Print Timestamped Messages
}
