% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_test_metrics}
\alias{process_test_metrics}
\title{Process Test Metrics from Results List: converts the nested list of test metrics into a long-format data frame}
\usage{
process_test_metrics(results_models, classifiers)
}
\description{
Process Test Metrics from Results List: converts the nested list of test metrics into a long-format data frame
}
\keyword{internal}
