% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_input_classifier}
\alias{rename_input_classifier}
\title{Rename Composite Input-Classifier Names}
\usage{
rename_input_classifier(x, sep = "_")
}
\arguments{
\item{x}{Character vector of composite names.}

\item{sep}{Character. The separator used in the original names (default "_").}
}
\value{
A character vector of pretty composite names.
}
\description{
Takes a name like "data_driven_rf" and converts it to "data-driven RF"
using the existing mapping functions in the package.
}
