% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{rf_model}
\alias{rf_model}
\title{Random Forest Model Training}
\usage{
rf_model(
  data,
  target_var,
  folds,
  predictors,
  metrics,
  control,
  directory_to_save = NULL
)
}
\arguments{
\item{data}{Data frame containing the training data.}

\item{target_var}{Character string of the target variable column name.}

\item{folds}{Resampling object (e.g., vfold_cv) for cross-validation.}

\item{predictors}{Character vector of predictor variable names.}

\item{metrics}{Metrics object for model evaluation.}

\item{control}{Control object for tuning and resampling.}

\item{directory_to_save}{Character string specifying the directory to save the model bundles. Default is NULL (do not save).}
}
\value{
A list containing the time taken for training, resampling results, and the fitted model object.
}
\description{
This function trains a Random Forest model using the \code{parsnip} and \code{workflows} packages.
It performs hyperparameter tuning using Bayesian optimization with the \code{tune} package,
evaluates the model using resampling, and saves the fitted model and resampling results as
bundles in a specified directory.
}
