% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{run_aggregation_pipeline}
\alias{run_aggregation_pipeline}
\title{Full Aggregation Pipeline (Example)}
\usage{
run_aggregation_pipeline(
  results_models_list,
  results_shap_list,
  threshold = 0.75,
  absolute = TRUE,
  candidate_results_dir = NULL
)
}
\arguments{
\item{results_models_list}{A named list of performance objects.}

\item{results_shap_list}{A named list of SHAP dataframes corresponding to the models.}

\item{threshold}{Numeric. Performance threshold for filtering models.}

\item{absolute}{Logical. Whether to use absolute values for the gene-sample matrix.}
}
\description{
This demonstrates how to use the individual steps to go from raw results to
a candidate gene list.
}
