% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{run_feature_selection}
\alias{run_feature_selection}
\title{Main Function for Feature Selection}
\usage{
run_feature_selection(
  procedure,
  expression_data,
  target_var,
  threshold_value = NULL,
  disease_code = NULL,
  dea_genes = NULL,
  mrmr_genes = NULL,
  alternative_genes = NULL,
  directory_to_load = NULL,
  directory_to_save = NULL,
  mrmr_path = "./mrmr"
)
}
\arguments{
\item{procedure}{The feature selection procedure to use.}

\item{expression_data}{The expression data frame.}

\item{target_var}{The target variable name.}

\item{threshold_value}{Threshold value for selection (if applicable).}

\item{disease_code}{Disease code for disease-related selections.}

\item{dea_genes}{List of differentially expressed genes (if applicable).}

\item{mrmr_genes}{List of mRMR selected genes (if applicable}

\item{alternative_genes}{List of alternative selected genes (if applicable).}

\item{directory_to_load}{Directory to load data from (if applicable).}

\item{directory_to_save}{Directory to save results to (if applicable).}

\item{mrmr_path}{Path to the mRMR executable.}
}
\value{
A character vector of selected gene names based on the specified procedure.
}
\description{
This function serves as the main entry point for performing feature selection based on
various procedures. It takes in the necessary parameters and calls the appropriate helper
functions to obtain the selected gene lists according to the specified procedure.
}
