% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust_ranking.R
\name{run_mclust_analysis}
\alias{run_mclust_analysis}
\title{Run Mclust Analysis for Gene Ranking}
\usage{
run_mclust_analysis(data, column = "aggregated_max", G = 1:9)
}
\arguments{
\item{data}{Dataframe. The gene summary output from aggregation.}

\item{column}{Character. The column name to analyze (default "aggregated_max").}

\item{G}{Numeric vector. The number of clusters to test (default 1:9).}
}
\value{
A list containing the Mclust model, BIC, ICL, used data, and column name.
}
\description{
This function fits Gaussian Mixture Models to the aggregated SHAP values to identify
underlying groups of genes.
}
