% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_helper_multi_rda}
\alias{save_helper_multi_rda}
\title{Internal helper to save multiple objects into one .Rda file}
\usage{
save_helper_multi_rda(obj_list, name_list, filename, directory, subfolder = "")
}
\description{
Internal helper to save multiple objects into one .Rda file
}
\keyword{internal}
