% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_helper_rda}
\alias{save_helper_rda}
\title{Internal helper to handle directory creation and saving .Rda files}
\usage{
save_helper_rda(data_obj, obj_name, filename, directory, subfolder = "")
}
\description{
Internal helper to handle directory creation and saving .Rda files
}
\keyword{internal}
