% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_helper_rds}
\alias{save_helper_rds}
\title{Internal helper to save .rds files}
\usage{
save_helper_rds(data, filename, directory, subfolder = "", verbose = TRUE)
}
\description{
Internal helper to save .rds files
}
\keyword{internal}
