% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_helper_txt}
\alias{save_helper_txt}
\title{Internal helper to handle directory creation and saving text files}
\usage{
save_helper_txt(data, filename, directory, subfolder = "")
}
\description{
Internal helper to handle directory creation and saving text files
}
\keyword{internal}
