% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust_ranking.R
\name{select_genes_mclust}
\alias{select_genes_mclust}
\title{Select Candidate Genes based on Mclust Clusters}
\usage{
select_genes_mclust(mclust_list, sel_groups)
}
\arguments{
\item{mclust_list}{List. Output from \code{run_mclust_analysis}.}

\item{sel_groups}{Numeric vector. The cluster IDs chosen as "high SHAP".}
}
\value{
A sorted dataframe of selected genes.
}
\description{
This function identifies candidate genes based on the clusters defined by the Mclust model.
It determines a threshold based on the minimum score of the selected clusters and filters the genes
accordingly. The resulting candidate genes are sorted by their SHAP scores and returned as a dataframe.
The function also prints the defined threshold and the number of selected genes to the console.
}
