% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_cross_validation_metrics}
\alias{process_cross_validation_metrics}
\title{Process Cross-Validation Metrics from Results List
converts the nested list of cross-validation metrics into a long-format data frame}
\usage{
process_cross_validation_metrics(results_models, classifiers)
}
\description{
Process Cross-Validation Metrics from Results List
converts the nested list of cross-validation metrics into a long-format data frame
}
\keyword{external}
